requirements.txt
django==5.1.1
djangorestframework==3.15.2
pytest-django==4.9.0
model-bakery==1.19.5
Faker==30.8.1
pyotp==2.9.0
django-extensions==3.2.3
numpy==2.2.0
pandas==2.2.3
pytz==2024.2
Pillow==11.0.0